<feat #130> 피드백 반영
