+++ src/main/java/com/prography/minari/admin/controller/AdminController.java
@@ -5,11 +5,10 @@ import com.prography.minari.admin.controller.dto.RewardForceRequest;
 import com.prography.minari.admin.controller.dto.RewardForceV2Request;
 import com.prography.minari.admin.service.AdminRewardService;
 import com.prography.minari.common.response.CommonResponse;
-import com.prography.minari.user.entity.User;
+import com.prography.minari.common.util.JwtUtil;
 import com.prography.minari.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -18,6 +17,7 @@ import org.springframework.web.bind.annotation.*;
 public class AdminController implements AdminApiDocs {
     private final AdminRewardService adminRewardService;
     private final UserService userService;
+    private final JwtUtil jwtUtil;
 
     @PostMapping("/payment/force")
     public ResponseEntity forceCharge(@RequestBody RewardForceRequest request) {
@@ -38,6 +38,18 @@ public class AdminController implements AdminApiDocs {
                 request.getMemo(),
                 request.getProductId());
         return ResponseEntity.ok(CommonResponse.ok());
+    }
+
+    @GetMapping("/token/expired")
+    public ResponseEntity<CommonResponse>  createExpiredToken(@RequestParam("userId") String userId) {
+        String expiredToken = jwtUtil.createExpiredToken(userId);
+        return ResponseEntity.ok(CommonResponse.success(expiredToken));
+    }
+
+    @DeleteMapping("/users/me")
+    public ResponseEntity<CommonResponse> deleteUser(@RequestParam("userId") Long userId) {
+        userService.deleteAdmin(userId);
+        return ResponseEntity.ok(CommonResponse.success("[ADMIN] 계정 즉시 삭제"));
     }
 
 }
